apps/api/pyproject.toml
[project]
name = "interior-vision-api"
version = "0.0.0"
description = "Interior Vision — FastAPI sidecar"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.34",
  "pydantic-settings>=2.7",
  "structlog>=25.1",
  "fal-client>=0.10",
  "Pillow>=10.0",
  "python-multipart>=0.0.9",
  "tenacity>=9.0",
]

[dependency-groups]
dev = [
  "httpx>=0.28",
  "pyinstaller>=6.12",
  "pytest>=8.3",
  "pytest-asyncio>=0.25",
  "pytest-cov>=6.0",
  "ruff>=0.11",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=term-missing --cov-fail-under=80"
testpaths = ["tests"]
markers = [
  "live: requires live external services and FAL_KEY env var (deselect with '-m not live')",
]
